requirements.txt
requests
aiohttp
tqdm
psycopg2-binary
lxml
prometheus-client